requirements - Copy.txt
  pandas
  scikit-learn
  xgboost
  mlflow
  uvicorn # For FastAPI
  fastapi # For API serving
#  faker # For data simulation
  joblib
  google-cloud-storage
  pydantic # For FastAPI's BaseModel